# test/Sample.Minimal.Tests/snapshots/FoundationTests.OpenApiDocument.verified.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sample.Minimal | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/launch-records": {
      "get": {
        "tags": [
          "LaunchRecordEndpoints"
        ],
        "operationId": "ListLaunchRecords",
        "parameters": [
          {
            "name": "rocketType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/NullableOfRocketType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaunchRecordModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "LaunchRecordEndpoints"
        ],
        "operationId": "CreateLaunchRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLaunchRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "location": {
                "description": "The location of the entity that was created",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLaunchRecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/launch-records/{id}": {
      "get": {
        "tags": [
          "LaunchRecordEndpoints"
        ],
        "operationId": "GetLaunchRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaunchRecordModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "LaunchRecordEndpoints"
        ],
        "operationId": "EditLaunchRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/LaunchRecordId2"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditLaunchRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaunchRecordModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "tags": [
          "LaunchRecordEndpoints"
        ],
        "operationId": "PatchLaunchRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/LaunchRecordId2"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditLaunchRecordPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaunchRecordModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "LaunchRecordEndpoints"
        ],
        "operationId": "DeleteLaunchRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/LaunchRecordId2"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/rockets": {
      "get": {
        "tags": [
          "RocketEndpoints"
        ],
        "parameters": [
          {
            "name": "rocketType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/NullableOfRocketType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RocketModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RocketEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRocketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "location": {
                "description": "The location of the entity that was created",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRocketResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/rockets/{id}": {
      "get": {
        "tags": [
          "RocketEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RocketId2"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RocketModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "RocketEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RocketId2"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRocketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RocketModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "tags": [
          "RocketEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RocketId2"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRocketPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RocketModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "RocketEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RocketId2"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssignedOfdouble": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          },
          "hasValue": {
            "type": "boolean"
          }
        }
      },
      "AssignedOfInstant": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/RocketId"
          },
          "hasValue": {
            "type": "boolean"
          }
        }
      },
      "AssignedOfNullableOfInstant": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/RocketId"
          },
          "hasValue": {
            "type": "boolean"
          }
        }
      },
      "AssignedOfRocketId": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/RocketId"
          },
          "hasValue": {
            "type": "boolean"
          }
        }
      },
      "AssignedOfRocketType": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/RocketType"
          },
          "hasValue": {
            "type": "boolean"
          }
        }
      },
      "AssignedOfstring": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          },
          "hasValue": {
            "type": "boolean"
          }
        }
      },
      "CreateLaunchRecordRequest": {
        "required": [
          "partner",
          "rocketId",
          "payload",
          "scheduledLaunchDate"
        ],
        "type": "object",
        "properties": {
          "rocketId": {
            "$ref": "#/components/schemas/Instant"
          },
          "partner": {
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "payload": {
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "payloadWeightKg": {
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "actualLaunchDate": {
            "$ref": "#/components/schemas/Instant"
          },
          "scheduledLaunchDate": {
            "$ref": "#/components/schemas/Instant"
          }
        }
      },
      "CreateLaunchRecordResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/LaunchRecordId"
          }
        }
      },
      "CreateRocketRequest": {
        "required": [
          "serialNumber"
        ],
        "type": "object",
        "properties": {
          "serialNumber": {
            "maxLength": 30,
            "minLength": 10,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/RocketType"
          }
        }
      },
      "CreateRocketResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RocketId"
          }
        }
      },
      "EditLaunchRecordPatchRequest": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RocketId"
          },
          "partner": {
            "$ref": "#/components/schemas/AssignedOfstring"
          },
          "payload": {
            "$ref": "#/components/schemas/AssignedOfstring"
          },
          "payloadWeightKg": {
            "$ref": "#/components/schemas/AssignedOfdouble"
          },
          "actualLaunchDate": {
            "$ref": "#/components/schemas/AssignedOfNullableOfInstant"
          },
          "scheduledLaunchDate": {
            "$ref": "#/components/schemas/AssignedOfInstant"
          },
          "rocketId": {
            "$ref": "#/components/schemas/AssignedOfRocketId"
          }
        }
      },
      "EditLaunchRecordRequest": {
        "required": [
          "id",
          "partner",
          "rocketId",
          "payload",
          "scheduledLaunchDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RocketId"
          },
          "partner": {
            "minLength": 1,
            "type": "string"
          },
          "payload": {
            "minLength": 1,
            "type": "string"
          },
          "payloadWeightKg": {
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "actualLaunchDate": {
            "$ref": "#/components/schemas/RocketId"
          },
          "scheduledLaunchDate": {
            "$ref": "#/components/schemas/RocketId"
          },
          "rocketId": {
            "$ref": "#/components/schemas/RocketId"
          }
        }
      },
      "EditRocketPatchRequest": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RocketId"
          },
          "serialNumber": {
            "$ref": "#/components/schemas/AssignedOfstring"
          },
          "type": {
            "$ref": "#/components/schemas/AssignedOfRocketType"
          }
        }
      },
      "EditRocketRequest": {
        "required": [
          "id",
          "type",
          "serialNumber"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RocketId"
          },
          "serialNumber": {
            "maxLength": 30,
            "minLength": 10,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/RocketType"
          }
        }
      },
      "Instant": { },
      "LaunchRecordId": { },
      "LaunchRecordId2": {
        "type": "string",
        "format": "uuid"
      },
      "LaunchRecordModel": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Instant"
          },
          "partner": {
            "type": "string"
          },
          "payload": {
            "type": "string"
          },
          "payloadWeightKg": {
            "type": "integer",
            "format": "int64"
          },
          "actualLaunchDate": {
            "$ref": "#/components/schemas/Instant"
          },
          "scheduledLaunchDate": {
            "$ref": "#/components/schemas/Instant"
          },
          "rocketSerialNumber": {
            "type": "string"
          },
          "rocketType": {
            "$ref": "#/components/schemas/RocketType"
          }
        }
      },
      "NullableOfRocketType": {
        "enum": [
          "falcon9",
          "falconHeavy",
          "atlasV",
          null
        ]
      },
      "RocketId": { },
      "RocketId2": {
        "type": "string",
        "format": "uuid"
      },
      "RocketModel": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RocketId"
          },
          "sn": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/RocketType"
          }
        }
      },
      "RocketType": {
        "enum": [
          "falcon9",
          "falconHeavy",
          "atlasV"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "LaunchRecordEndpoints"
    },
    {
      "name": "RocketEndpoints"
    }
  ]
}